+++ src/components/manage/Controlpanels/AddonsControlpanel.jsx
@@ -27,7 +27,7 @@ import {
   upgradeAddon,
 } from '@plone/volto/actions';
 import { Helmet } from '@plone/volto/helpers';
-import { Icon, Toolbar } from '@plone/volto/components';
+import { Icon, Toolbar, Error } from '@plone/volto/components';
 import circleBottomSVG from '@plone/volto/icons/circle-bottom.svg';
 import circleTopSVG from '@plone/volto/icons/circle-top.svg';
 import backSVG from '@plone/volto/icons/back.svg';
@@ -230,6 +230,9 @@ class AddonsControlpanel extends Component {
    * @returns {string} Markup for the component.
    */
   render() {
+    if (this.props.error) {
+      return <Error error={this.props.error} />;
+    }
     return (
       <Container id="page-addons" className="controlpanel-addons">
         <Helmet title="Addons" />
@@ -464,6 +467,7 @@ export default compose(
       installedAddons: state.addons.installedAddons,
       availableAddons: state.addons.availableAddons,
       upgradableAddons: state.addons.upgradableAddons,
+      error: state.addons.error,
       pathname: props.location.pathname,
     }),
     { installAddon, listAddons, uninstallAddon, upgradeAddon },